lan-photo-share/requirements.txt
quart==0.19.6
hypercorn==0.16.0
SQLAlchemy[asyncio]==2.0.31
aiofiles==24.1.0
aiosqlite==0.20.0
Pillow==10.4.0